Staged dashboard for other models.
We first started by creating the class LeaderInformation. This could use a better name and perhaps we can use it elsewhere, lets discuss that. Next we put some sample data directly in the controller for the test. Finally the view had to use the LeaderInformation via an each block, inserting the elements into table rows.